kimi-convergence-loop/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kimi-convergence-loop"
version = "0.1.0"
description = "Self-healing convergence loop for iterative system improvement"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Kimi Ecosystem"}
]
keywords = ["convergence", "self-healing", "pipeline", "security", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "types-pyyaml>=6.0",
]

[project.scripts]
kimi-converge = "kimi_convergence_loop.cli:cli"

[project.urls]
Homepage = "https://github.com/kimi-ecosystem/kimi-convergence-loop"
Repository = "https://github.com/kimi-ecosystem/kimi-convergence-loop"

[tool.hatch.build.targets.wheel]
packages = ["src/kimi_convergence_loop"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
